Made the spread parameter to spawn blood a float (instead of int).
git-svn-id: file:///home/tjaden/{repo}/redstone/trunk@43 ec8ac367-a008-0410-9901-8f3136df4969

+++ src/bindgen.lua
@@ -384,7 +384,7 @@ generate {
     args	= {{ Float, "x" },
 		   { Float, "y" },
 		   { Int, "nparticles" },
-		   { Int, "spread" }},
+		   { Float, "spread" }},
     ret		= { Int, "ret", "$ < 0" }
 }
